Agents: Agent data and types in one place
We created a new folder called `agents`, analogous to the `models`
directory. This way, agent types can be reused by different models.
